Create dec4 codeing
Changes user input entirely.  created like 5 functions and two methods. the functions relate to the user input changes, essentially adding more options for learning. the methods integrated the symptom based search which works by comparing two sets, scoring them based on how similar they are, creating a dictionary with the symptoms as keys and the relating scores as values. the user input symptoms are being compared to each of the  diagnoses' symptoms.